convert to in-app message as much as possible
